refactor: 경로 설정 및 Atom 컴포넌트 사용 수정
- 상대경로 → alias 경로로 변경
- Atom 컴포넌트 리팩토링에 따른 Atom 컴포넌트 사용 수정
