+++ src/pages/Board.tsx
@@ -1,11 +1,5 @@
-import Button from "../components/Button";
+import { Button } from '@/components';
 
-const Board = () => {
-  return (
-    <Button additionalProps="This is button" onClick={() => console.log("Button Clicked")}>
-      글쓰기
-    </Button>
-  );
-};
-
-export default Board;
+export default function Board() {
+  return <Button onClick={() => console.log('Button Clicked')}>글쓰기</Button>;
+}
